% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_details.R
\name{analysis}
\alias{analysis}
\title{Get analysis details}
\usage{
analysis(
  type = c("sample", "analysis"),
  name = NULL,
  id = NULL,
  start_date = NULL,
  end_date = NULL,
  duration = NULL,
  exclude = NULL,
  view = NULL,
  tidy = TRUE
)
}
\arguments{
\item{type}{character. Either "sample" (the default) or "analysis"}

\item{name}{character. See details.}

\item{id}{character. Optional. ID of analysis.}

\item{start_date, end_date}{character. Optional. YYYY-MM-DD format.}

\item{duration}{Optional. Parameter to filter the results by days/months.
Valid value starts with "-" follow by a numeric value and ending with "d"
or "m"}

\item{exclude}{Optional. Can be "reports", "unfilteredvariants",
"filteredvariants", or NULL. If NULL (default), downloads all.}

\item{view}{Optional. Can be "summary", which will include a list of all
analyses without links to the analyses (which can be faster to call)}

\item{tidy}{boolean. Ignored if \code{view} is NULL. Should the results be
presented as a tidy \code{tibble}?}
}
\value{
A JSON-like list or, if \code{view = "summary"} and \code{tidy = TRUE}, a
\code{tibble}.
}
\description{
Get analysis details
}
