% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download an IonReporter file}
\usage{
download(file_path, type = NULL, dest = NULL)
}
\arguments{
\item{file_path}{Path where the file is stored on the server. See details.}

\item{type}{Format of file. By default, file will be downloaded enclosed in a
.zip}

\item{dest}{Location where the file should be saved. If NULL, will be
downloaded to tempfile.}
}
\value{
Path to downloaded files
}
\description{
Download an IonReporter file
}
