% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bam.R
\name{download_bam}
\alias{download_bam}
\title{Download a BAM file}
\usage{
download_bam(file_path, type = NULL)
}
\arguments{
\item{file_path}{character. Typically the output of
\code{get_associated_bamfiles}.}

\item{type}{character. Honestly no idea what this does. If NULL (default),
downloads a zip. Other options are "" and "pdf"}
}
\description{
In a typical workflow, this will be used with \code{get_associated_bamfiles} to
obtain the \code{file_path} argument
}
