% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_associated_bamfiles.R
\name{get_associated_bamfiles}
\alias{get_associated_bamfiles}
\title{Get BAM files associated with samples or analyses}
\usage{
get_associated_bamfiles(name, type = c("analysis", "sample"))
}
\arguments{
\item{name}{character vector of either sample or analysis names}

\item{type}{character, either "sample" or "analysis"}
}
\description{
Get BAM files associated with samples or analyses
}
\details{
You can get a list of analysis names using the \code{list_analyses}
function. It's unclear if there's a way to get a list of sample names from
the API.
}
