% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcf.R
\name{getvcf}
\alias{getvcf}
\alias{get_vcf}
\alias{download_vcf}
\title{Get a link to a VCF file}
\usage{
getvcf(
  type = c("sample", "analysis"),
  name,
  id = NULL,
  start_date = NULL,
  end_date = NULL,
  duration = NULL,
  exclude = NULL
)

get_vcf(...)

download_vcf(dest = NULL, ...)
}
\arguments{
\item{type}{character. Either "sample" (the default) or "analysis"}

\item{name}{character. See details.}

\item{id}{character. Optional. ID of analysis.}

\item{start_date, end_date}{character. Optional. YYYY-MM-DD format.}

\item{duration}{Optional. Parameter to filter the results by days/months.
Valid value starts with "-" follow by a numeric value and ending with "d"
or "m"}

\item{exclude}{Optional. Can be "unfilteredvariants", "filteredvariants", or
NULL. If NULL (default), downloads both.}

\item{...}{Additional arguments passed to \code{getvcf}.}

\item{dest}{Where the file should be saved. Should generally be a .zip. If
null, will be saved as a tempfile with a \code{.zip} extension}
}
\value{
\itemize{
\item \code{download_vcf}: The path to the downloaded file
\item \code{getvcf}: A JSON-like list containing the link to download the file, the
name, and the id of the analysis that the sample was in
}
}
\description{
If you want to download a VCF to a file, \code{download_vcf} is a better option.
}
\details{
\code{download_vcf} provides a slightly more ergonomic way to directly download
files, rather than just get links to them.

Sample names are the names in the 'Sample' field in under the
'Samples' tab on IonReporter. They are also under the 'samples' column in
the output of \code{get_analysis_details}, \emph{not} the 'name' column.

Analysis names can also be used to select a vcf to download/get links to
download. They exist in the 'name' column in the output of
\code{get_analysis_details}

Although the original API documentation seems to imply that the \code{name}
argument isn't needed if \code{type == "analysis"}, it appears that it is still
needed.

All other arguments - except \code{exclude} - are further filters of the hits
where \code{type} and \code{name} match.

I wish I could give a better explanation of "duration" but honestly the API
documentation was sparse and gave no examples
}
