% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_setup.R
\name{key_setup}
\alias{key_setup}
\title{Helper to set up IonReporter envvar}
\usage{
key_setup()
}
\value{
boolean. Has key been set? Will return FALSE after setup since
session hasn't been restarted
}
\description{
Helper to set up IonReporter envvar
}
